# DRAFT: C#-Performance:  Use a custom static lookup to handle InvokeGodotClassMethod and HasGodotClassMethod