--- a/modules/mono/glue/GodotSharp/GodotSharp/Core/NativeInterop/InteropStructs.cs
+++ b/modules/mono/glue/GodotSharp/GodotSharp/Core/NativeInterop/InteropStructs.cs
@@ -533,7 +533,7 @@ namespace Godot.NativeInterop
         internal readonly unsafe godot_string_name* GetUnsafeAddress()
             => (godot_string_name*)Unsafe.AsPointer(ref Unsafe.AsRef(in _data));
 
-        private IntPtr _data;
+        internal IntPtr _data;
 
         public void Dispose()
         {
@@ -584,7 +584,7 @@ namespace Godot.NativeInterop
         [StructLayout(LayoutKind.Sequential)]
         internal struct movable
         {
-            private IntPtr _data;
+            internal IntPtr _data;
 
             public static unsafe explicit operator movable(in godot_string_name value)
                 => *(movable*)CustomUnsafe.AsPointer(ref CustomUnsafe.AsRef(value));