@roookeee 2024-03-24
I could use some help with fixing the pipeline because the diff isn't really helping me:
![image](https://github.com/godotengine/godot/assets/1199562/f0895ca8-2f2a-4169-945c-1d5d61288352)
I checked for LF / CLRF issues but the whole file is LF like other places.