@paulloz 2024-03-24
I expect this is a bom issue, double-check the file encoding between utf8 and utf8-bom.