@roookeee 2024-03-24
Thanks, I was lazy it seems :D I fixed the BOM issue and will move on to the other broken pipeline steps. May I get a workflow approval for future pipeline runs? Is this approval manual for every push I do for draft MRs?