@roookeee 2024-03-26
I gave this one more round of polish, names are more in line with standard Godot wording an I refactored the `ScriptMethod` delegate to not the return `bool` as that should be handled by `InvokeGodotClassMethod` - that seemed to have gained me 2-3ms (~5%) locally, but I won't edit the main post because I feel like that could be noise from less stuff running in parallel on my machine. The remaining TODOs are about discussions to be had.

P.S: If someone with some spare time and an actual Godot C# project stumbles upon this PR, I would love to hear if this PR breaks or improves stuff for you!