@roookeee 2024-03-28
It seems like I miscalculated how much the release template performance would compare to the "Editor -> Run Project": The gains in the release template are small, around 10% with this PR, nearly 20% when adding the other referenced PR.

I added the new numbers to the main post, heres the updates test project which shows the process time in a `Label`
[VariantOptimizationTestProject_New.zip](https://github.com/godotengine/godot/files/14795742/VariantOptimizationTestProject_New.zip)