@roookeee 2024-03-28
And now I wonder if my test project gets kind of dead code eliminated in release builds because `_Process` is a nop, need to rework it.