@roookeee 2025-01-29
I am rescinding this this PR as there seems to be little interest on the outlined approach. Understandable with a new C# support on the horizon :)